Allow TOTP codes that start with zero
ngx_snprintf()'s "%i" format specifier does not work with
variable-length fields ("*"), which prevented potential codes
from being zero-padded on the left.  This then prevented them
from matching properly zero-padded codes from the user.

Additionally, the "%i" format specifier is for ngx_int_t.
This commit also switches to "%uD", which is uint32_t,
which is the declared return type for the argument (which is
the ngx_http_auth_totp_algorithm_hotp() function).